refactor(tvc-utils): move fake manifest helpers into a fake module
